requirements.txt
Flask
Pillow
numpy
pandas
matplotlib
scikit-image
scipy
Werkzeug
pytest